The Semifields Module

​

Semifields — Module

    module Semifields: A module to capture the concept of a semifield and provide most notorious examples.

​

Module Index

​

Semifields.plusTwo

Detailed API

​

Semifields.plusTwo — Method

plusTwo(x)

Sum the numeric "2" to whatever it receives as input

A more detailed explanation can go here, although I guess it is not needed in this case

Arguments

x: The amount to which we want to add 2

Notes

Notes can go here

Examples

julia> five = plusTwo(3)
5